make long poll longer
change from 9 to 50 secs, the maximum.  this reduces network activity but does mean that things "wait" as this is not async. Don't really need the bot to be very responsive except for respending to messages so this has no impact. if we had lots of users it would be a problem but then we don't